NOSQL_TASK/Flask_mongo/requirements.txt
streamlit
pymongo
python-dotenv
flask
